test(landing): extract shared sole-door scan machinery out of a test module
Mission charter-sole-door-bypass-closure-01KZ3WAA, FR-043 (close defect
classes by construction) / adversarial-review gate hardening. Gates 1
(agent-profile repository) and 2 (raw DoctrineService) shared their
qualname-resolution AST machinery by importing it from Gate 1's own
test_charter_sole_door_agent_profile_repository.py module — library code
living inside a pytest-collected test_ module, which src/ cannot import and
which drifts (Gate 4/5 each rolled independent, already-diverging copies of
iter_source_files/parent_map/repo-root derivation instead of reusing it).

Moves REPO_ROOT, SRC_ROOT, iter_source_files, rel_to_repo, module_name_for,
parent_map, enclosing_scope, _scope_chain, _dotted_name, _lookup_module,
resolve_canonical, FileScan, ConstructionSite, ScanResult,
scan_file_constructions, scan_constructions, structurally_exempt,
resolve_exclusion_keys, assert_rationales_are_substantive, and scratch_scan
into a new non-test library module, tests/architectural/_sole_door_scan.py,
mirroring the existing tests/architectural/_ratchet_keys.py convention. Gates
1 and 2 now both import from it; no other gate is repointed yet (Gate 3
already imported only the small subset it needs from Gate 1 and gets
repointed alongside its own fix; Gates 4 and 5 keep their private copies for
now and are repointed in their own fold commits below).

Pure move: no behavior change. All 39 Gate 1 + Gate 2 tests still pass
unchanged.
